Move ICE fallback prompt to time of placing / answering calls
This moves the ICE fallback prompt out of session startup and instead it will
now appear contextually when your either place a call with no ICE server from
the homeserver or a call fails (in either direction).

Fixes https://github.com/vector-im/riot-web/issues/10546
